Fixing random key creation during accessing List items
When List.__getitem__ is called, it results in the creation of a random key.  This is because RedisCollection._create_new is called without passing the key argument.  So it defaults to the creation of a random key.
This can be fixed by passing the key of the calling List object
Fixes #23